area-tmx: Maps are editable in Tiled without editing
Until now, one had to mess with the file paths for tileset and tileset
image sources to get them to open in Tiled.  After finishes ones
changes in Tiled, one undid the path changes for Tsunagari.  This was a
tedious process.

Tsunagari now speaks Tiled paths for tileset and tileset image sources.