Change player holding key from CTRL to SHIFT
Previously, the user could hold down left CTRL while hitting an arrow
key which would cause the player to turn to face the direction indicated
by the arrow key but not start moving. This commit changes the key to
SHIFT (either left or right) to fix conflicts with the Mac OS X window
manager.