Fix area color overlays
Area color overlays were not implemented correctly. They took the screen
size in physical pixels and tried to draw an overlay of that size in
virtual pixels starting at the top-left corner of the map. If the
viewport was sufficiently far away from the origin, the overlay would
not cover the screen.

Fix the problem by drawing the overlay before we enter area-space.